sing_app/sing_app/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
alembic==1.7.5
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-decouple==3.5
python-dotenv==0.19.2
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1s
Flask-Login==0.5.0
email-validator==1.1.3
Flask-Mail==0.9.1
pytest-flask==1.2.0
# pgadmin4==6.5